Updated Sign up and Login pages to have consistent css with the rest of our app. Renamed LoginSignup to Signup. Removed LoginSignup.html as it was not needed. Updated icons to be consistent with our other icons. Fixed error messages to match for both sign up and login pages
